+++ tools/__init__.py
@@ -5,6 +5,7 @@ from .video_tools import GenerateVideoTool
 from .video_task_tools import QueryVideoTaskTool, DownloadVideoTool
 from .music_tools import GenerateMusicTool
 from .music_cover_tools import MusicCoverTool
+from .background_task_tools import QueryBackgroundTaskTool
 from .search_tools import WebSearchTool
 from .vision_tools import DescribeImageTool
 from .quota_tools import CheckQuotaTool
@@ -17,6 +18,7 @@ __all__ = [
     "DownloadVideoTool",
     "GenerateMusicTool",
     "MusicCoverTool",
+    "QueryBackgroundTaskTool",
     "WebSearchTool",
     "DescribeImageTool",
     "CheckQuotaTool",